+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Fixed
 
+- When `igraph_adjlist_init()` was called on a directed graph that had mutual edges, but no multi-edges, using `mode = IGRAPH_ALL` and `multiple = IGRAPH_NO_MULTIPLE`, it would incorrectly set the `IGRAPH_PROP_HAS_MULTI` cache entry to true.
 - `igraph_read_graph_graphml()` imported the edge IDs incorrectly from GraphML files. This bug was introduced in 1.0.0. Thanks to @zxawry for the fix (PR #2894)!
 
 ## [1.0.1]